stock-chatbot-main/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
requests
pinecone
huggingface-hub
google-generativeai
python-dotenv
